Add Gazebo plugin for setting USV pose and speed
The plugin makes it possible to set the position (x, y, z), orientation
(roll, pitch, yaw), and forward speed of the USV directly. This is done
by sending a std_msgs::Float64MultiArray on the topic
"gazebo_usv_direct_control_plugin/usv_data" where the fields are in the
order mentioned above.

The plugin is disabled by default, but there is an argument for enabling
it in otter.launch in otter_gazebo package.